+++ internal/server/model.go
@@ -1,0 +1,13 @@
+package server
+
+type cat struct {
+	Name  string `json:"name" xml:"name"`
+	Color string `json:"color" xml:"color"`
+}
+
+type methodResponse map[string]response
+
+type response struct {
+	status int
+	body   any
+}