+++ internal/server/server.go
@@ -26,7 +26,10 @@ func New(opts Options) *Server {
 
 	handler := newHandler(ch)
 	mux := http.NewServeMux()
-	mux.HandleFunc("/", handler.handle)
+	mux.HandleFunc("/~/success/", handler.handleSuccess)
+	mux.HandleFunc("/~/client-errors/", handler.handleClientErrors)
+	mux.HandleFunc("/~/server-errors/", handler.handleServerErrors)
+	mux.HandleFunc("/", handler.handleDefault)
 
 	s := &http.Server{
 		Addr:    fmt.Sprintf(":%d", opts.Port),
@@ -63,4 +66,25 @@ func (s *Server) Close() error {
 		s.done <- true
 	}()
 	return s.server.Close()
+}
+
+func ListRoutes() {
+	fmt.Println("Success routes:")
+	for route, methods := range successResponses {
+		for method, resp := range methods {
+			fmt.Printf("  %s %s (status %d)\n", method, route, resp.status)
+		}
+	}
+
+	// FIXME: sort by status
+	fmt.Println("\nClient Errors (accepts any method)")
+	for route, status := range clientErrors {
+		fmt.Printf("  %s %d\n", route, status)
+	}
+
+	// FIXME: sort by status
+	fmt.Println("\nServer Errors (accepts any method)")
+	for route, status := range serverErrors {
+		fmt.Printf("  %s %d\n", route, status)
+	}
 }